% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_accuracy}
\alias{bc_accuracy}
\title{Calculate breath counting accuracy}
\usage{
bc_accuracy(df, y)
}
\arguments{
\item{df}{Data frame containing breath counting data}

\item{y}{Tibble Grouping variable (see group_modify())}
}
\value{
Data frame
}
\description{
Calculate breath counting accuracy
(Levinson, Stoll, Kindy, Merry, & Davidson, 2014)
}
\keyword{breath}
\keyword{counting}
\keyword{meditation}
\keyword{mindfulness}
