% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_all_accuracy}
\alias{bc_all_accuracy}
\title{Calculate breath counting accuracy for a group of participants}
\usage{
bc_all_accuracy(df, c = "subject")
}
\arguments{
\item{df}{Data frame. Breath counting accuracy for participants}

\item{c}{String. Column to group by}
}
\value{
Data frame
}
\description{
Calculate breath counting accuracy for a group of participants.
}
