% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_process_eprime}
\alias{bc_process_eprime}
\title{Process ePrime breath counting data files}
\usage{
bc_process_eprime(path)
}
\arguments{
\item{path}{Path to ePrime data files}
}
\value{
Data frame
}
\description{
Process ePrime breath counting data files.
}
