% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_process_eprime_file}
\alias{bc_process_eprime_file}
\title{Process ePrime breath counting data file}
\usage{
bc_process_eprime_file(path)
}
\arguments{
\item{path}{Path to ePrime data file}
}
\value{
Data frame
}
\description{
Process ePrime breath counting data file.
(Levinson, Stoll, Kindy, Merry, & Davidson, 2014)
}
\keyword{ePrime}
