% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_process_expfactory_file}
\alias{bc_process_expfactory_file}
\title{Process Experiment Factory breath counting task data file}
\usage{
bc_process_expfactory_file(path, p, json = TRUE)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}

\item{json}{Boolean indicating whether data format is JSON (TRUE), default or CSV (FALSE)}
}
\value{
Data frame
}
\description{
\code{process_breath_counting()} processes data files generated by the Breath Counting Task
(\url{https://expfactory-experiments.github.io/breath-counting-task/}).
}
\keyword{breath}
\keyword{expfactory}
\keyword{mindfulness}
