% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{breath_counting}
\alias{breath_counting}
\title{Convert ePrime or Experiment Factory breath counting data to data frame}
\usage{
breath_counting(
  data_path,
  type = "expfactory",
  format = "json",
  participants,
  t
)
}
\arguments{
\item{data_path}{Path to data files}

\item{type}{String ('expfactory' or 'eprime, default='expfactory')}

\item{format}{String ('csv' or 'json', default='json')}

\item{participants}{Data frame}

\item{t}{Numeric Time point}
}
\value{
Data frame
}
\description{
Convert ePrime or Experiment Factory breath counting data to data frame.
}
