% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{breath_counting_accuracy}
\alias{breath_counting_accuracy}
\title{Breath Counting Accuracy}
\usage{
breath_counting_accuracy(df)
}
\arguments{
\item{df}{Breath counting data frame}
}
\description{
Process ePrime breath counting data
(Levinson, Stoll, Kindy, Merry, & Davidson, 2014)
}
\keyword{breath}
\keyword{counting}
\keyword{meditation}
\keyword{mindfulness}
