% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{breath_counting_to_csv}
\alias{breath_counting_to_csv}
\title{Convert ePrime or Experiment Factory breath counting data to CSV}
\usage{
breath_counting_to_csv(data_path, type = "expfactory", format = "json",
  participants, t, exclude)
}
\arguments{
\item{data_path}{Path to data files}

\item{type}{String ('expfactory' or 'eprime, default='expfactory')}

\item{format}{String ('csv' or 'json', default='json')}

\item{participants}{Data frame}

\item{t}{Numeric Time point}

\item{exclude}{Participants to exclude}
}
\description{
Convert ePrime or Experiment Factory breath counting data to CSV
}
