% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{eprime_breath_counting_to_csv}
\alias{eprime_breath_counting_to_csv}
\title{Convert ePrime breath counting data to CSV}
\usage{
eprime_breath_counting_to_csv(path)
}
\arguments{
\item{path}{Path to ePrime data files}
}
\description{
Convert ePrime breath counting data to CSV
}
