% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\docType{package}
\name{expfactory.breathcounting}
\alias{expfactory.breathcounting}
\title{expfactory.breathcounting: A package for processing breath counting data.}
\description{
This package processes data generated from a breath counting task
which is a behavioural measure of mindfulness. The package can analyse data from the
original ePrime task, and a version which runs in a web browser. The latter version
was designed to be used within The Experiment Factory framework.
}
\section{expfactory.breathcounting functions}{

...
}

\references{
Levinson, D. B., Stoll, E. L., Kindy, S. D., Merry, H. L., and Davidson, R. J. (2014). A mind you can count on: validating breath counting as a behavioral measure of
mindfulness. Frontiers in Psychology, 5.
}
