% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{n_breath_counting_accuracy}
\alias{n_breath_counting_accuracy}
\title{Group Breath Counting Accuracy}
\usage{
n_breath_counting_accuracy(df, c = "subject")
}
\arguments{
\item{df}{Data frame. Breath counting accuracy for n participants}

\item{c}{String. Column to group by}
}
\description{
Process ePrime breath counting data for n participants
}
