% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.breathcounting.R
\name{bc_process_expfactory}
\alias{bc_process_expfactory}
\title{Process Experiment Factory breath counting data files}
\usage{
bc_process_expfactory(path, t)
}
\arguments{
\item{path}{Path to data files}

\item{t}{Time point}
}
\value{
Data frame
}
\description{
Process Experiment Factory breath counting data files.
}
